+++ react/IconCaretLeft.js
@@ -1,0 +1,3 @@
+import CaretLeft from './components/icon/CaretLeft'
+
+export default CaretLeft