+++ react/IconCaretRight.js
@@ -1,0 +1,3 @@
+import CaretRight from './components/icon/CaretRight'
+
+export default CaretRight