+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- **CaretLeft** Export CaretLeft component as an app to render.
+- **CaretRight** Export CaretRight component as an app to render.
+
 ## [4.2.1] - 2018-06-14
 
-- **Checkbox** Export Checkbox component as an app to render. 
+- **Checkbox** Export Checkbox component as an app to render.
 
 ## [4.2.0] - 2018-06-12